etl-appfigures-reviews/requirements.txt
requests_oauthlib
google-cloud-secretmanager
pandas
pandas-gbq
flask
google-cloud-logging
google-cloud-bigquery